+++ efile_app/efile/static/js/upload-handler.js
@@ -92,7 +92,7 @@ class UploadHandler {
                     method: 'POST',
                     headers: {
                         'Content-Type': 'application/json',
-                        'X-CSRFToken': this.getCSRFToken()
+                        'X-CSRFToken': apiUtils.getCSRFToken()
                     },
                     body: caseFormData
                 });
@@ -115,7 +115,7 @@ class UploadHandler {
                 method: 'POST',
                 headers: {
                     'Content-Type': 'application/json',
-                    'X-CSRFToken': this.getCSRFToken()
+                    'X-CSRFToken': apiUtils.getCSRFToken()
                 },
                 body: JSON.stringify(uploadData)
             });
@@ -181,7 +181,7 @@ class UploadHandler {
                 method: 'POST',
                 headers: {
                     'Content-Type': 'application/json',
-                    'X-CSRFToken': this.getCSRFToken()
+                    'X-CSRFToken': apiUtils.getCSRFToken()
                 },
                 body: JSON.stringify(uploadData)
             });
@@ -204,10 +204,6 @@ class UploadHandler {
             console.error('Error saving files to session:', error);
             throw error;
         }
-    }
-
-    getCSRFToken() {
-        return document.querySelector('[name=csrfmiddlewaretoken]')?.value || '';
     }
 
     setupEventListeners() {
@@ -523,16 +519,6 @@ class UploadHandler {
         checkGlobalData();
     }
 
-    getCSRFToken() {
-        return (
-            document.querySelector("[name=csrfmiddlewaretoken]")?.value ||
-            document
-                .querySelector('meta[name="csrf-token"]')
-                ?.getAttribute("content") ||
-            ""
-        );
-    }
-
     setupSupportingDocumentCascading(index) {
         const filingTypeSelect = document.getElementById(`supportingFilingType${index}`);
         const documentTypeSelect = document.getElementById(`supportingDocumentType${index}`);
@@ -602,7 +588,7 @@ class UploadHandler {
                 method: 'POST',
                 body: formData,
                 headers: {
-                    'X-CSRFToken': this.getCSRFToken()
+                    'X-CSRFToken': apiUtils.getCSRFToken()
                 }
             });
