+++ efile_app/efile/templates/efile/components/profile_header.html
@@ -77,7 +77,7 @@
         method: 'POST',
         headers: {
           'Content-Type': 'application/json',
-          'X-CSRFToken': getCookie('csrftoken')
+          'X-CSRFToken': window.apiUtils.getCSRFToken()
         },
         body: JSON.stringify({ jurisdiction: jurisdiction })
       });
@@ -117,21 +117,6 @@
     setTimeout(() => {
       alertDiv.remove();
     }, 3000);
-  }
-
-  function getCookie(name) {
-    let cookieValue = null;
-    if (document.cookie && document.cookie !== '') {
-      const cookies = document.cookie.split(';');
-      for (let i = 0; i < cookies.length; i++) {
-        const cookie = cookies[i].trim();
-        if (cookie.substring(0, name.length + 1) === (name + '=')) {
-          cookieValue = decodeURIComponent(cookie.substring(name.length + 1));
-          break;
-        }
-      }
-    }
-    return cookieValue;
   }
 
   function logout() {
